refactor: divide MyProfile in multiple pages
Dividido o conteúdo de MyProfile em dois componentes vue como página separadas a fim de modularizar, separar lógicas independentes e rotear apropriadamente cada parte.
